back/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
gspread>=5.12.4
oauth2client>=4.1.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
qrcode[pil]>=7.4.2
Pillow>=10.1.0
reportlab>=4.0.7
slowapi>=0.1.9
requests>=2.31.0
boto3>=1.34.0
pypdf>=3.17.0